% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phase3-Create_Decision_Data_Table.R
\name{CreateLPDT}
\alias{CreateLPDT}
\title{Create Linear Program Data Table}
\usage{
CreateLPDT(db)
}
\arguments{
\item{db}{}
}
\value{
data.table
}
\description{
Create Linear Program Data Table
}
