% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phase3-GetTargetPopDensity.R
\name{GetTargetPopDensity}
\alias{GetTargetPopDensity}
\title{Get Target Population Density}
\usage{
GetTargetPopDensity()
}
\value{
data.table containing list of targets and zctas
}
\description{
Reads in shapefiles describing the geometry of the region, and some data
files that describe the population of each geometry. Finds the geometry
containing each target's spatialPoint. Note that not all spatialPoints will
belong to a geometry.
}
