% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phase3-InitializeData.R
\name{pKB}
\alias{pKB}
\title{Cumulative Probability of Kill}
\usage{
pKB(mer, cep, tgt.n = 1, tgt.r = 0, tot.r = 0, m = 1)
}
\arguments{
\item{mer}{Mean Effective Radius}

\item{cep}{Circular Error Probable}

\item{tgt.n}{number of targets (assumed to be 1 unless otherwise specified)}

\item{tgt.r}{radius of target (assumed to be 0 unless otherwise specified)}

\item{tot.r}{total radius containing randomly distributed identical targets (assumed to be 0 unless otherwise specified)}

\item{m}{the number of munitions of the given type fired on the target}
}
\value{
probability of total kill
}
\description{
Cumulative Probability of Kill
}
