% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phase3-InitializeData.R
\name{qKB}
\alias{qKB}
\title{Quantity of Munitions to Achieve pKB}
\usage{
qKB(p, mer, cep, tgt.n = 1, tgt.r = 0, tot.r = 0)
}
\arguments{
\item{p}{probability of total kill}

\item{mer}{Mean Effective Radius of munition against target}

\item{cep}{Circular Error Probable of munition}

\item{tgt.n}{number of targets}

\item{tgt.r}{radius of each target}

\item{tot.r}{radius of area containing the identical, randomly distributed targets.}
}
\value{
a number of shots that would need to be fired to achieve the desired pKB
}
\description{
Quantity of Munitions to Achieve pKB
}
